backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Google Cloud Storage
google-cloud-storage==2.10.0

# Environment and validation
python-dotenv==1.0.0
pydantic[email]==2.5.0
email-validator==2.1.0

# CORS and middleware
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2 